backend_flask/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.1
pytesseract==0.3.10
ultralytics==8.0.196
pillow==10.0.1
python-dateutil==2.8.2
openpyxl==3.1.2 